Move debug functions to its own file
