+++ odoo/debug.go
@@ -1,0 +1,60 @@
+package odoo
+
+import (
+	"bytes"
+	"fmt"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"regexp"
+
+	"github.com/go-logr/logr"
+)
+
+type debugTransport struct {
+	pwRe *regexp.Regexp
+}
+
+func newDebugTransport() *debugTransport {
+	return &debugTransport{
+		pwRe: regexp.MustCompile(`("password":\s?").+("[,}])`),
+	}
+}
+
+// UseDebugLogger sets the http.Transport field of the internal http client with a transport implementation that logs the raw contents of requests and responses.
+// The logger is retrieved from the request's context via logr.FromContextOrDiscard.
+// The log level used is '2'.
+// Any "password":"..." byte content is replaced with a placeholder to avoid leaking credentials.
+// Still, this should not be called in production as other sensitive information might be leaked.
+// This method is meant to be called before any requests are made (for example after setting up the Client).
+func (c Client) UseDebugLogger(enabled bool) {
+	if enabled {
+		c.http.Transport = newDebugTransport()
+	}
+}
+
+// RoundTrip implements http.RoundTripper.
+func (t *debugTransport) RoundTrip(r *http.Request) (*http.Response, error) {
+	logger := logr.FromContextOrDiscard(r.Context())
+	if r.Body != nil {
+		reqBody, _ := r.GetBody()
+		defer reqBody.Close()
+		buf, _ := ioutil.ReadAll(reqBody)
+		buf = t.pwRe.ReplaceAll(buf, []byte(`$1[confidential]$2`))
+		logger.V(2).Info(fmt.Sprintf("%s %s ---> %s", r.Method, r.URL.Path, string(buf)))
+	}
+
+	res, err := http.DefaultTransport.RoundTrip(r)
+	if err != nil {
+		return nil, err
+	}
+
+	if res.Body != nil {
+		defer res.Body.Close()
+		buf, _ := ioutil.ReadAll(res.Body)
+		logger.V(2).Info(fmt.Sprintf("%s %s <--- %s", r.Method, r.URL.Path, string(buf)))
+		res.Body = io.NopCloser(bytes.NewReader(buf))
+	}
+
+	return res, err
+}
